refactor(printRcingResult): 함수를 더 작은 단위로 분리
