+++ src/App.js
@@ -2,7 +2,7 @@ import { Console } from '@woowacourse/mission-utils';
 import { inputCarNames } from './functions/inputCarNames.js';
 import { inputAttemptsNumber } from './functions/inputAttempsNumber.js';
 import { calculateCarRacing } from './functions/calculateCarRacing.js';
-import { printRacingResult } from './functions/printRacingResult.js';
+import { printRacingResult } from './functions/printRacingResultFunctions/printRacingResult.js';
 import { calculateWinner } from './functions/calculateWinner.js';
 import { printWhoIsWinner } from './functions/printWhoIsWinner.js';
 import { validateCarNames } from './errors/allErrorHandling.js';
