+++ src/functions/printRacingResultFunctions/printRacingResult.js
@@ -1,23 +1,18 @@
 import { Console } from '@woowacourse/mission-utils';
+import { initializeCumulativeResults } from './initializeCumulativeResults.js';
+import { printSingleAttemptResult } from './printSingleAttemptResult.js';
 
 /**
+ * 전체 경주 결과를 출력
  * @param {Array} resultsPerAttempt - 각 시도별 경주 결과 배열
  * @param {Array} carNames - 자동차 이름 배열
  */
 export function printRacingResult(resultsPerAttempt, carNames) {
     Console.print('\n실행 결과');
 
-    const cumulativeResults = carNames.reduce((acc, car) => {
-        acc[car] = '';
-
-        return acc;
-    }, {});
+    const cumulativeResults = initializeCumulativeResults(carNames);
 
     resultsPerAttempt.forEach((attemptResult) => {
-        carNames.forEach((car) => {
-            cumulativeResults[car] += attemptResult[car]; // 누적 결과 갱신하기
-            Console.print(`${car} : ${cumulativeResults[car]}`);
-        });
-        Console.print('');
+        printSingleAttemptResult(attemptResult, carNames, cumulativeResults);
     });
 }
